Additional serial speeds
I can consistently get my boards to download at 921,600. This patch adds
in all the standard to 2,000,000. The symptoms of going too fast are
that it never initially connects. I have never had a corrupted download.

The new baud rates are commented out for w32.

http://code.google.com/p/stm32flash/issues/detail?id=17

Tested-by: Reuben Dowle <reuben.dowle@navico.com>

[Tormod]
Limit some speeds to GNU (Linux) since they are not supported on
e.g. Mac OS X. Thanks to Mike Smith for pointing out this in
http://code.google.com/p/stm32flash/issues/detail?id=38

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>