New reset code common for Cortex-M0, Cortex-M3, Cortex-M4
http://code.google.com/p/stm32flash/issues/detail?id=34

[Tormod]
Add references and more comments to reset code

I also found this link which is publicly available but the link looks
volatile and the web page has "superseded" painted on it:

http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.ddi0337g/BHCFCFDB.html

[Tormod]
Fix assembled binary reset code

Although the assembly in the initial patch was right, the actual assembled
byte code was wrongly aligned, and most likely caused a hardfault.

The new code is generated with:

echo "   ldr r1, =0xe000ed0c
         ldr r2, =0x05fa0004
         str r2, [r1]
      endless: b endless " |
  arm-none-eabi-as -mcpu=cortex-m3 -mthumb

and has been verified "live" with a JTAG debugging adapter.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>