Remove old reset code cross-compiling infrastructure
The new reset code is short and relatively simple and there is no need
to keep the whole infrastructure for cross-compiling a full executable
including startup code and reset vector table and generate C code to
embed it.

A small difference in the actual reset code is that the new code does
not preserve the PRIGROUP bits when writing to the AIRCR register:

         SCB->AIRCR = (NVIC_AIRCR_VECTKEY |
-                      (SCB->AIRCR & 0x700) |
                       (1<<NVIC_SYSRESETREQ));
         while(1);

The NVIC_SystemReset() in ST's core_cm3.h CMSIS library header
also preserves the PRIGROUP and also adds a "dsb" command but I think
none of this is needed.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>