stm32: bootloader command list
AN3155 reports a static reply list for GET command.
But list can change across devices, e.g. if one or more
commands are not supported by a custom bootloader.

Change the code to handle dynamically the length of
the reply and to manage unsupported commands.

Moreover, this patch is required to support STM32W
bootloader that does not provide "write unprotect".

Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>