Rewrite and extend memory address conversion
The conversion of addresses from command line parameters to internals
was spread around the code.
Consolidate it in a single part of code and handle all addresses, either
flash, ram and registers.
Registers can be read/write in blocks of max 4 bytes.

Detect command line parameter "-e 0" and use it as flag to skip
flash erase.

Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>