serial_posix: be more accepting of device paths
On Linux, it is possible to refer to USB serial devices with durable
paths such as /dev/serial/by-path/pci-0000:00:14.0-usb-0:1:1.0-port0 or
/dev/serial/by-id/usb-FTDI_FT232R_USB_UART_A700abCD-if00-port0.  These
paths are themselves symlinks to /dev/tty*.

I have also found it convenient to place symlinks in my build
directories, so that the file 'PORT' names the appropriate device for
the project.

This patch enables both usages, by first using the standard POSIX API
realpath(3) to resolve symbolic links, and only then checks that the
device name starts with /dev/tty.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>