main: collaps few variables in a single enum
main() implements only one actions at a time (or read or write
or ...).
We were using few variables to flag which action to take.
Also, we had incomplete tests to avoid the command line to list
two or more actions.

Replace variables with a single enum and enforce the tests on
the command line.

Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>