dev_table: Use array for page size of flash
Some devices like F2, F4 and F7 have flash split in pages having
different size.
Replace the scalar number fl_ps with a zero terminated array.
Change type of fl_ps to accommodate the size of big pages.
Specify 1 page per sector for F2, F4 and F7.
Modify the code to use the first element of fl_ps array in
place of the scalar number.

This patch does not change the overall behavior of the code.

Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>