Use Thumb mode jump address for GO command
As discussed in the ST forums [1], the bootloader requires
an odd (Thumb mode) address for the GO command. If an even address
is used, the Cortex M does something like a software reset, but this
is not generated by the write to AICR in stm_reset_code, but by an
exception due to the even address which is not compatible with
Thumb mode. ARM specifies 'Hard Fault' for this case:
http://infocenter.arm.com/help/
index.jsp?topic=/com.arm.doc.dui0497a/BABDCDHA.html

[1] https://my.st.com/public/STe2ecommunities/mcu/Lists/cortex_mx_stm32/System%20boot%20mode%20and%20Go%20command%20STM32f103rbt6

Signed-off-by: Ernst Schwab <eschwab@online.de>