Allow software reset generation combined with init string
The stm32_reset_device() software reset generation function is only
called if the -R option is specified. However, if an initialization
string (-i) was added, stm32_reset_device() was not called, and
gpio_bl_exit() was used instead, even if it didn't contain an exit
string.

This change will instead allow the combination of a GPIO initialization
sequence and a software reset, as long as there is no exit sequence.

Also an empty exit sequence will prevent the software reset.

Examples:
  -R -i rts:-rts   init string, exit string, no sw reset
  -R -i rts:       init string, empty exit string, no sw reset
  -R -i rts        init string, no exit string, sw reset

Signed-off-by: Ernst Schwab <eschwab@online.de>