Allow read/write to full option byte area and system memory
Set the end address to the end of the memory area in question.
Specifying shorter length is possible using -S address:length

Note that some hardware only let you read from the beginning of the
option byte area (or write the complete area in one operation).

Thanks to Edvar Klakken for reporting and Maurits van der
Hoorn for suggesting and testing an initial patch.

https://sourceforge.net/p/stm32flash/tickets/93/

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>